Update README.md
A python script to run the popular game pong written in python. 
It is a two player game, player one uses the w and s keys to move their bar up and down, and player two uses the up and down arrow keys to move their bar up and down to play the game. Each time a point is scored, the game will pause for 3 seconds, and once a player gets 15 points, the game ends.